requirements.txt
fastapi==0.109.2
pytest==8.0.1
pytest-asyncio==0.23.5
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic-settings==2.2.0